fix(chore): moved DotPushPublishEnvironmentsResolver to Ui lib
