+++ core-web/apps/dotcms-ui/src/app/portlets/dot-edit-page/dot-edit-page.module.ts
@@ -15,8 +15,7 @@ import {
     DotExperimentsConfigResolver,
     DotExperimentsService
 } from '@dotcms/portlets/dot-experiments/data-access';
-import { DotEnterpriseLicenseResolver } from '@dotcms/ui';
-import { DotPushPublishEnvironmentsResolver } from '@portlets/shared/resolvers';
+import { DotEnterpriseLicenseResolver, DotPushPublishEnvironmentsResolver } from '@dotcms/ui';
 import { DotFeatureFlagResolver } from '@portlets/shared/resolvers/dot-feature-flag-resolver.service';
 import { DotDirectivesModule } from '@shared/dot-directives.module';
