+++ core-web/apps/dotcms-ui/src/app/portlets/shared/resolvers/index.ts
@@ -1,2 +1,1 @@
 export * from './dot-feature-flag-resolver.service';
-export * from './dot-push-publish-enviroments-resolver.service';
