+++ core-web/libs/portlets/dot-experiments/portlet/src/lib/dot-experiments-portlet-routes.ts
@@ -2,8 +2,7 @@ import { Routes } from '@angular/router';
 
 import { ExperimentsConfigProperties } from '@dotcms/dotcms-models';
 import { DotExperimentsConfigResolver } from '@dotcms/portlets/dot-experiments/data-access';
-import { DotEnterpriseLicenseResolver } from '@dotcms/ui';
-import { DotPushPublishEnvironmentsResolver } from '@portlets/shared/resolvers';
+import { DotEnterpriseLicenseResolver, DotPushPublishEnvironmentsResolver } from '@dotcms/ui';
 
 import { DotExperimentsAnalyticAppMisconfigurationComponent } from './dot-experiments-analytic-app-misconfiguration/dot-experiments-analytic-app-misconfiguration.component';
 import { DotExperimentsConfigurationComponent } from './dot-experiments-configuration/dot-experiments-configuration.component';
