+++ core-web/libs/ui/src/index.ts
@@ -54,6 +54,7 @@ export * from './lib/pipes/dot-safe-html/dot-safe-html.pipe';
 // Resolvers
 export * from './lib/resolvers/dot-portlet-have-license.resolver';
 export * from './lib/resolvers/dot-enterprise-license-resolver.service';
+export * from './lib/resolvers/dot-push-publish-enviroments-resolver.service';
 
 // Validators
 export * from './lib/validators/dotValidators';
