+++ core-web/libs/utils-testing/src/index.ts
@@ -45,3 +45,4 @@ export * from './lib/dot-current-user.mock';
 export * from './lib/split-button.mock';
 export * from './lib/seo-mock';
 export * from './lib/dot-message-mock.pipe';
+export * from './lib/push-publish.mock';
