+++ core-web/libs/utils-testing/src/lib/push-publish.mock.ts
@@ -1,0 +1,19 @@
+import { of as observableOf } from 'rxjs';
+
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class MockPushPublishService {
+    getEnvironments() {
+        return observableOf([
+            {
+                id: '123',
+                name: 'Environment 1'
+            },
+            {
+                id: '456',
+                name: 'Environment 2'
+            }
+        ]);
+    }
+}
